Fix circular import in vespa_loader by moving ingest_document_service to separate module
